fix: creating prewiev photo from video

+++ src/Domain/Common/Interfaces/Services/IFfmpegService.cs
@@ -4,6 +4,6 @@ public interface IFfmpegService
 {
     Task<string> UploadGeneratedVideos(string videoPath, CancellationToken token = default);
     Task<string> RunProcess(string args, string fileName = "ffmpeg", CancellationToken token = default);
-    Task<int> GetVideoDuration(string videoPath, CancellationToken token = default);
-    Task<string> GetPhotoFromVideo(string videoPath, string outputPath = "/storage/images/previewPhoto.jpeg", int timeSeconds = 5, CancellationToken token = default);
+    Task<double> GetVideoDuration(string videoPath, CancellationToken token = default);
+    Task<string> GetPhotoFromVideo(string videoPath, string outputPath = "/storage/images/previewPhoto.jpeg", double timeSeconds = 5, CancellationToken token = default);
 }
